Import ::Unicode.const_missing to rex/text/unicode
This delta was generated from the SVIT fork's version of rex/text.
Git blame shows this diff to be part of 0e3b1585 on 20130430.
Other changes have been omitted as they do not pertain to upstream
imports of the fork's functional changes.

Commit message indicates this to be part of the Powershell cleanup
effort during one of the merges upstream, and designed to deal with
parsing foreign PSH scripts containing aberrant characters.